Conflicts:
	DjVuDoc.cpp
	DjVuView.cpp
	Global.cpp
	Languages/Russian/WinDjViewRU.vcproj
	MainFrm.cpp
	MyTheme.h
	WinDjView.sln
	WinDjView.vcproj

+++ ProgressDlg.cpp
@@ -1,5 +1,5 @@
 //	WinDjView
-//	Copyright (C) 2004-2009 Andrew Zhezherun
+//	Copyright (C) 2004-2012 Andrew Zhezherun
 //
 //	This program is free software; you can redistribute it and/or modify
 //	it under the terms of the GNU General Public License as published by
@@ -15,8 +15,6 @@
 //	with this program; if not, write to the Free Software Foundation, Inc.,
 //	51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 //	http://www.gnu.org/copyleft/gpl.html
-
-// $Id: ProgressDlg.cpp,v 1.17 2009/08/07 19:35:40 zhezherun Exp $
 
 #include "stdafx.h"
 #include "WinDjView.h"
@@ -122,6 +120,6 @@ void CProgressDlg::OnOK()
 
 LRESULT CProgressDlg::OnEndDialog(WPARAM wParam, LPARAM lParam)
 {
-	EndDialog(wParam);
+	EndDialog((int)wParam);
 	return 0;
 }
